refactor(app): app 명령어로 app create 를 사용하도록 수정
동일 소스로 app 명령어와 app create 명령어를 실행하도록 변경
